refactor(titlebar): separar lógica de controles de ventana (SRP)
